requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
chex==0.1.89
click==8.2.1
cloudpickle==3.1.1
colour==0.1.5
contourpy==1.3.2
control==0.10.2
cycler==0.12.1
decorator==5.2.1
dm-tree==0.1.9
einops==0.8.1
equinox==0.13.0
etils==1.12.2
exceptiongroup==1.3.0
executing==2.2.0
flax==0.10.7
fonttools==4.58.5
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.9.0
humanize==4.12.3
idna==3.10
importlib_resources==6.5.2
ipdb==0.13.13
ipython==8.37.0
jax==0.6.2
jax-cuda12-pjrt==0.6.2
jax-cuda12-plugin==0.6.2
jaxlib==0.6.2
jaxtyping==0.3.2
jedi==0.19.2
jraph==0.0.6.dev0
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.4
ml_dtypes==0.5.1
msgpack==1.1.1
mujoco==3.3.4
mujoco-mjx==3.3.4
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.3.9
opencv-python==4.11.0.86
opt_einsum==3.4.0
optax==0.2.5
orbax-checkpoint==0.11.19
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyOpenGL==3.1.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.32.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
tensorflow-probability==0.25.0
tensorstore==0.1.76
tomli==2.2.1
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.9
trimesh==4.7.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wadler_lindig==0.1.7
wandb==0.21.0
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.23.0
